// src/app/interfaces/product.interface.ts
// Generated by https://quicktype.io

export interface Product {
  id:          number;
  title:       string;
  price:       number;
  description: string;
  images:      string[];
  creationAt:  string;
  updatedAt:   string;
  category:    Category;
  amount: number;
}

export interface Category {
  id:         number;
  name:       string;
  image:      string;
  creationAt: string;
  updatedAt:  string;
}